% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dados}
\alias{dados}
\title{Subcategorias da CID-10}
\format{
A data frame com 12451 linhas e 3 variáveis:
\describe{
  \item{SUBCAT}{Códigos da CID-10.}
  \item{DESCRICAO}{Descrição dos códigoc da CID-10.}
  \item{code_class}{Códigos de classificadores das causas de óbitos evitáveis na infância.}
}
}
\source{
<https://www.github.com/ronaldoalves-ms/stopdeath>
}
\usage{
dados
}
\description{
Data frame com códigos de subcategorias da CID-10, enriquecido com a variável "code_class".
}
\keyword{datasets}
