% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stop_birthdefect.R
\name{stop_birthdefect}
\alias{stop_birthdefect}
\title{Classificação de Anomalias Congênitas Prioritárias}
\usage{
stop_birthdefect(dados, var_nome)
}
\arguments{
\item{dados}{Um data frame contendo os dados de interesse.}

\item{var_nome}{Um vetor de caracteres contendo os códigos da CID-10.}
}
\value{
A função retorna um data frame enriquecido com variáveis classificatórias.
}
\description{
\code{class_birthdefect} é uma função de identificação e classificação de códigos da CID-10 para fins de vigilância de anomalias congênitas no Brasil.
}
\examples{
library(stopdeath)
dados <- stop_birthdefect(dados, dados$var_nome)

}
\references{
Cardoso-dos-Santos AC, Medeiros-de-Souza AC, Bremm JM, Alves RFS, et al. Lista de anomalias congênitas prioritárias para vigilância no âmbito do Sistema de Informações sobre Nascidos Vivos do Brasil. Epidemiologia e Serviços de Saúde, 30(1):e2020835, 2021.\url{https://www.scielo.br/scielo.php?script=sci_arttext&pid=S2237-96222021000100600&tlng=pt}
}
\author{
Ronaldo Fernandes Santos Alves. Fundação Oswaldo Cruz, Instituto de Comunicação e Informação Científica e Tecnológica em Saúde.
}
