% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stop_death.R
\name{stop_death}
\alias{stopdeath}
\title{Codificação de Causas de Óbitos Evitáveis na Infância}
\usage{
stop_death(dados, var_nome)
}
\arguments{
\item{dados}{Um data frame contendo os dados de interesse.}

\item{var_nome}{Um vetor de caracteres contendo os códigos da CID-10.}
}
\value{
A função retorna um data frame enriquecido com a variável "code_class".
}
\description{
\code{stop_death} é uma função de codificação de causas de óbitos evitáveis em crianças menores de 5 anos de idade. A função cria um vetor de códigos identificadores (code_class), cuja estrutura compreende variados níveis hierárquicos de classificação de causas de óbitos evitáveis.
}
\examples{
library(stop_death)
dados <- stop_death(dados, dados$var_nome)

}
\references{
Malta DC, Duarte EC, Almeida MF, et al. Lista de causas de mortes evitáveis por intervenções do Sistema Único de Saúde do Brasil. Epidemiologia e Serviços de Saúde, 16(4):233-244, out-dez 2007.\url{http://scielo.iec.gov.br/pdf/ess/v16n4/v16n4a02.pdf}

Malta DC, Sardinha LMV, Moura L, et al. Atualização da lista de causas de mortes evitáveis por intervenções do Sistema Único de Saúde do Brasil. Epidemiologia e Serviços de Saúde, 19(2):173-176, abr-jun 2010.\url{http://scielo.iec.gov.br/pdf/ess/v19n2/v19n2a10.pdf}
}
\author{
Ronaldo Fernandes Santos Alves. Fundação Oswaldo Cruz, Instituto de Comunicação e Informação Científica e Tecnológica em Saúde.
}
