% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dados}
\alias{dados}
\title{Codificação de Causas de Óbitos Evitáveis na Infância}
\format{
Um data frame com 12451 linhas e 1 variável:
\describe{
  \item{var_nome}{Códigos de subcategorias da CID-10}
}
}
\source{
<https://www.github.com/ronaldoalves-ms/stopdeath>
}
\usage{
dados
}
\description{
Um data frame com códigos de subcategorias da CID-10.
}
\keyword{datasets}
